requirements.txt
openai>=1.3.0
cryptography>=42.0
pyyaml>=6.0
numpy>=1.26
pytest>=8.0
bandit>=1.7.7
detect-secrets>=1.4.0
black>=24.3
flake8>=7.0
fastapi>=0.100.0
tinydb>=4.7
opentelemetry-api>=1.0
